% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstreeOutput.R
\name{jstreeOutput}
\alias{jstreeOutput}
\title{Shiny jstree output}
\usage{
jstreeOutput(outputId, checkbox = FALSE, search = FALSE,
  search_placeholder = "")
}
\arguments{
\item{outputId}{jstree identification}

\item{checkbox}{TRUE/FALSE}

\item{search}{TRUE/FALSE}

\item{search_placeholder}{string}
}
\description{
Shiny jstree output
}
