% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{+.plugin}
\alias{+.plugin}
\alias{-.plugin}
\title{Add/remove child plugins from plugin object}
\usage{
\method{+}{plugin}(e1, e2)

\method{-}{plugin}(e1, e2)
}
\arguments{
\item{e1}{plugin object}

\item{e2}{child to add/remove}
}
\value{
plugin object
}
\description{
+ will add/replace a plugin child to a plugin object, 
- will remove a child from a plugin object
}
\details{
DETAILS
}
\examples{
p <- plugin(name = 'search',
attr = list(case_sensitive = FALSE,
            show_only_matches = FALSE))

p

p1 <- plugin(name = 'dnd',
             attr = list(copy = FALSE))

p1

p2 <- p + p1

p2

p2 - p + p1
}
