% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{plugin}
\alias{plugin}
\title{jsTree plugin builder api}
\usage{
plugin(name, attr)
}
\arguments{
\item{name}{character, name of plugin}

\item{attr}{list, names of attributes relevant to the plugin}
}
\value{
list
}
\description{
api to iteratively create jsTree plugins
}
\details{
plugins and their characteristics can be found at \url{https://www.jstree.com/api/}
}
\examples{
p <- plugin(name = 'search',
attr = list(case_sensitive = FALSE,
            show_only_matches = FALSE))

p

p1 <- plugin(name = 'dnd',
             attr = list(copy = FALSE))

p1


p2 <- p + p1

p2

p2 + plugin(name = 'dnd',
             attr = list(copy = TRUE))

p2 - p + p1
}
