% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{print.plugin}
\alias{print.plugin}
\title{Print method for plugin objects}
\usage{
\method{print}{plugin}(x, ...)
}
\arguments{
\item{x}{plugin object}

\item{\dots}{unused}
}
\description{
Prints a plugin object in json form
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
