% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setNode}
\alias{setNode}
\title{set jstree node}
\usage{
setNode(x, children = list(), icon = NULL, stopened = FALSE,
  stselected = FALSE, stdisabled = FALSE)
}
\arguments{
\item{x}{string}

\item{children}{setNode object, list os setNodes or empty list}

\item{icon}{Font Awesome icon}

\item{stopened}{TRUE/FALSE}

\item{stselected}{TRUE/FALSE}

\item{stdisabled}{TRUE/FALSE}
}
\description{
set jstree node
}
