% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_bird-data.R
\docType{data}
\name{state_bird}
\alias{state_bird}
\title{Character vector of state birds}
\format{
A character vector of length 50
}
\source{
\url{https://state.1keydata.com/state-birds.php}
}
\usage{
state_bird
}
\description{
Character vector of state birds
}
\keyword{datasets}
