% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states-data.R
\docType{data}
\name{states}
\alias{states}
\title{State data}
\format{A data frame with 400 rows and 5 variables:
\describe{
  \item{\code{state.region}}{factor State Region}
  \item{\code{state.division}}{factor State Sub Region}
  \item{\code{state.name}}{character State Name}
  \item{\code{variable}}{factor Characteristic}
  \item{\code{value}}{double Characteristic Value} 
}}
\usage{
states
}
\description{
state dataset melted into a data.frame
}
\keyword{datasets}
