% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsTree.R
\name{jsTree}
\alias{jsTree}
\title{Htmlwidget for the jsTree Javascript library}
\usage{
jsTree(obj, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{obj}{character, vector of directory tree}
}
\description{
Htmlwidget for the jsTree Javascript library
}
\examples{
jsTree(list.dirs())
if(dir.exists(.git)) jsTree(system('git ls-tree -r HEAD --name-only',intern=TRUE))
}
