% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsTree.R
\name{jsTree}
\alias{jsTree}
\title{Htmlwidget for the jsTree Javascript library}
\usage{
jsTree(obj, tooltips = NULL, nodestate = NULL, preview.search = NULL,
  remote_repo = NULL, remote_branch = "master", vcs = "github",
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{obj}{character, vector of directory tree}

\item{tooltips}{character, named vector of tooltips for elements in the tree, Default: NULL}

\item{nodestate}{boolean, vector the length of obj that initializes tree open to true values, Default: NULL}

\item{preview.search}{character, Search term to initialize to in the preview pane searchbox, Default: NULL}

\item{remote_repo}{character, remote user/repository, Default: NULL}

\item{remote_branch}{character, branch of remote_repo, Default: 'master'}

\item{vcs}{character, choose which version control system to attach (github, bitbucket, svn), Default: 'github'}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{The input slot that will be used to access the element.}
}
\description{
Htmlwidget for the jsTree Javascript library
}
\details{
if remote_repo is given a preview pane of a selected file from the tree will appear to the right of the tree.
preview.search is only relevant for vcs in (github,bitbucket) where file previewing is available
}
\examples{
if(interactive())
jsTree(list.files(full.names = TRUE,recursive = TRUE))
}
