% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsTree.R
\name{jsTree}
\alias{jsTree}
\title{Htmlwidget for the jsTree Javascript library}
\usage{
jsTree(obj, gh_repo = NULL, gh_branch = "master", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{obj}{character, vector of directory tree}

\item{gh_repo}{character, github user/repository, Default: NULL}

\item{gh_branch}{character, branch of gh_repo, Default: 'master'}
}
\description{
Htmlwidget for the jsTree Javascript library
}
\examples{
jsTree(list.dirs())
jsTree(ciderhouse::show_repo('tidyverse/purrr',showTree = FALSE))
}
