% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsTree.R
\name{jsTree}
\alias{jsTree}
\title{Htmlwidget for the jsTree Javascript library}
\usage{
jsTree(obj, tooltips = NULL, nodestate = NULL, remote_repo = NULL,
  remote_branch = "master", vcs = "github", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{obj}{character, vector of directory tree}

\item{tooltips}{character, named vector of tooltips for elements in the tree, Default: NULL}

\item{remote_repo}{character, remote user/repository, Default: NULL}

\item{remote_branch}{character, branch of remote_repo, Default: 'master'}

\item{vcs}{character, choose which version control system to attach (github, bitbucket, svn), Default: 'github'}
}
\description{
Htmlwidget for the jsTree Javascript library
}
\details{
if remote_repo is given a preview pane of a selected file from the tree will appear to the right of the tree.
}
\examples{
jsTree(list.files(full.names = TRUE,recursive = TRUE))
jsTree(vcs::ls_remote('tidyverse/reprex'),remote_repo = 'tidyverse/reprex')
}
