% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsTree.R
\name{jsTree}
\alias{jsTree}
\title{Htmlwidget for the jsTree Javascript library}
\usage{
jsTree(obj, gh_repo = NULL, gh_branch = "master", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{obj}{character, vector of directory tree}

\item{gh_repo}{character, github user/repository, Default: NULL}

\item{gh_branch}{character, branch of gh_repo, Default: 'master'}
}
\description{
Htmlwidget for the jsTree Javascript library
}
\details{
if gh_repo is given a preview pane of a selected file from the tree will appear to the right of the tree.
}
\examples{
jsTree(list.files(full.names = TRUE,recursive = TRUE))
jsTree(ciderhouse::ls_github('tidyverse/reprex',recursive = TRUE),gh_repo = 'tidyverse/reprex')
}
