% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{jsTree}
\alias{jsTree}
\title{R6 class for the jsTree Javascript library}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
jsTree
}
\description{
Htmlwidget for the jsTree Javascript library
}
\examples{
if(interactive()){

data(states)
data(state_bird)

x <- jsTree::jsTree$new()

# add data (can be data.frame or character)
 # each row of the data.frame columns are 
 # collapsed with x$sep to create a character vector
 
x$data <- states

x$data

# show the widget
 x$show()

# add tooltips
 x$tooltips <- state_bird
 x$show()

# add plugins
# search and checkbox are added as defaults
 x$active_plugins

 x$active_plugins.append(c('contextmenu','dnd','unique'))

 x$active_plugins

 x$show()

# add callbacks or options to a plugin
 x$plugins[['dnd']] <- list(is_draggable = FALSE)
 # x$plugins[['contextmenu']] <- list(items = JS('script here'))

x$plugins

x$show()

# change core options
 x$core_params <- list(multiple=FALSE)

# add inital nodestate with logical call that are applied to x$data
 x$nodestate(variable=='Area')

x$show()


}

}
\keyword{datasets}
